Update Final Project Help
The input isn't checking correctly against the list. When I type an element that is in the list, it isn't adding it to the cor_element_names list. 
Also, I'm having to input 6 names when I should only be inputting 5. The sixth also doesn't show on either of the lists.